+++ simulator/src/main/scala/textworldexpress/goldagent/MapReaderGoldAgent.scala
@@ -3,7 +3,7 @@ package textworldexpress.goldagent
 import textworldexpress.games.{ArithmeticGame, CoinGame, MapReaderGame, MapReaderRandomGame}
 import textworldexpress.objects.Room
 import textworldexpress.runtime.PythonInterface
-import textworldexpress.symbolicmodule.{ModuleCalc, ModuleNavigation}
+import textworldexpress.symbolicmodule.{ModuleCalc, ModuleKnowledgeBaseTWC, ModuleNavigation}
 
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
@@ -224,6 +224,13 @@ class MapReaderGoldAgentWithModule(interface:PythonInterface) {
 
   // Make the gold path, using the Calc module
   private def mkGoldPathMapReaderWithNavModule(r:Random): Boolean = {
+    // Ensure that we're running on the correct game type (MapReaderGame)
+    if (game == null) return false
+    // Ensure that the appropriate modules have been enabled
+    if (!interface.enabledModuleStrs.contains(ModuleCalc.MODULE_NAME)) {
+      println ("MapReaderGoldAgentWithModule: ERROR: " + ModuleNavigation.MODULE_NAME + " is not enabled -- can not generate gold path.")
+    }
+
     // Step 1: Read task description
     interface.step("task")
 
@@ -320,11 +327,6 @@ class MapReaderGoldAgentWithModule(interface:PythonInterface) {
   }
 
   def mkGoldPath(r:Random):(Boolean, Array[String]) = {
-    if (!interface.moduleInterface.getEnabledModuleNames().contains(ModuleNavigation.MODULE_NAME)) {
-      println ("ERROR: Navigation module (" + ModuleNavigation.MODULE_NAME + ") required to generate gold paths is not enabled.")
-      return (false, Array.empty[String])
-    }
-
     val success = this.mkGoldPathMapReaderWithNavModule(r)
     if (!success) return (false, Array.empty[String])
 
@@ -338,121 +340,3 @@ class MapReaderGoldAgentWithModule(interface:PythonInterface) {
 
 
 
-/*
- * Agent with module support (MapReaderRandom)
- */
-class MapReaderRandomGoldAgentWithModule(interface:PythonInterface) {
-  val game = interface.game match { case x:MapReaderRandomGame => x; case _=> null }
-
-  // Make the gold path, using the Calc module
-  private def mkGoldPathMapReaderWithNavModule(r:Random): Boolean = {
-    // Step 1: Read task description
-    interface.step("task")
-
-    // Step 2: Read map
-    interface.step("read map")
-
-    // Step 3: Start walking to location
-    if (!this.navigateToLocationUsingModule(game.endLocation)) {
-      //return false
-    }
-
-    // Step 4: Take coin
-    interface.step("take coin")
-
-    // Step 5: Return to start location
-    if (!this.navigateToLocationUsingModule(game.startLocation)) {
-      //return false
-    }
-
-    // Step 6: Put coin in box
-    interface.step("put coin in box")
-
-    return true
-
-  }
-
-  private def navigateToLocationUsingModule(endLocation:Room): Boolean = {
-    val MAX_ITER = 20
-    var numIter:Int = 0
-
-    while (game.agentLocation.name != endLocation.name) {
-      val stepResult = interface.step("next step to " + endLocation.name)
-      val obs = stepResult.observationStr
-
-      val fields = obs.split(":")
-      if (fields.length == 2) {
-        val locationStep = fields(1).trim()
-
-        // North
-        if ((game.agentLocation.locationNorth != null) && (game.agentLocation.locationNorth.name.toLowerCase == locationStep)) {
-          // Check to see if there is a door that needs to be opened first
-          if ((game.agentLocation.doorNorth != null) && (!game.agentLocation.doorNorth.isOpen)) {
-            interface.step("open door to north")
-          }
-          // Move north
-          interface.step("move north")
-
-          // South
-        } else if ((game.agentLocation.locationSouth != null) && (game.agentLocation.locationSouth.name.toLowerCase == locationStep)) {
-          // Check to see if there is a door that needs to be opened first
-          if ((game.agentLocation.doorSouth != null) && (!game.agentLocation.doorSouth.isOpen)) {
-            interface.step("open door to south")
-          }
-          interface.step("move south")
-
-          // East
-        } else if ((game.agentLocation.locationEast != null) && (game.agentLocation.locationEast.name.toLowerCase == locationStep)) {
-          // Check to see if there is a door that needs to be opened first
-          if ((game.agentLocation.doorEast != null) && (!game.agentLocation.doorEast.isOpen)) {
-            interface.step("open door to east")
-          }
-          interface.step("move east")
-
-          // West
-        } else if ((game.agentLocation.locationWest != null) && (game.agentLocation.locationWest.name.toLowerCase == locationStep)) {
-          // Check to see if there is a door that needs to be opened first
-          if ((game.agentLocation.doorWest != null) && (!game.agentLocation.doorWest.isOpen)) {
-            interface.step("open door to west")
-          }
-          interface.step("move west")
-        } else {
-          if (game.agentLocation.name.toLowerCase == locationStep) {
-            // Already at this location -- likely just the first or last step from the pathfinding algorithm.  Just ignore this one.
-          } else {
-            // If we reach here, the path must be incorrect, as the next location on the path isn't accessible from the current location.
-            println("ERROR: Unable to find location (" + locationStep + ")")
-            println("FROM Location: " + game.agentLocation.getDescription())
-            println("")
-            return false
-          }
-        }
-      }
-
-      numIter += 1
-      if (numIter > MAX_ITER) {
-        println ("ERROR: Timed out on navigating path (MAX_ITER = " + MAX_ITER + ")")
-        return false
-      }
-
-    }
-
-    // If we reach here, the navigation was successful
-    return true
-  }
-
-  def mkGoldPath(r:Random):(Boolean, Array[String]) = {
-    if (!interface.moduleInterface.getEnabledModuleNames().contains(ModuleNavigation.MODULE_NAME)) {
-      println ("ERROR: Navigation module (" + ModuleNavigation.MODULE_NAME + ") required to generate gold paths is not enabled.")
-      return (false, Array.empty[String])
-    }
-
-    val success = this.mkGoldPathMapReaderWithNavModule(r)
-    if (!success) return (false, Array.empty[String])
-
-    // Success
-    val path = interface.history.map(_.userInputStr).toArray.filter(_.length > 0)
-    return (true, path)
-  }
-
-}
